Switch to using 64 bit full words

+++ src/test.h
@@ -1,6 +1,6 @@
 /* asm test */
 
-unsigned testfun(unsigned);
+unsigned long testfun(unsigned long);
 
 /* Local Variables: */
 /* mode: c */
